requirements.txt
flask==3.0.3
requests==2.32.3
geopy==2.4.1
psycopg2==2.9.9
python-dotenv==1.0.1